% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carrots.R
\name{Carrots}
\alias{Carrots}
\title{Carrots R6 Class Definitions for testing}
\description{
Class in which every method has an equal number of package function references
}
\concept{Carrots}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Carrots-new}{\code{Carrots$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Carrots-new"></a>}}
\if{latex}{\out{\hypertarget{method-Carrots-new}{}}}
\subsection{Method \code{new()}}{
Initialize Carrots.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Carrots$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing
}
}
}
