% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carrots.R
\name{Carrots}
\alias{Carrots}
\title{Carrots}
\description{
Carrots
}
\concept{Carrots}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Carrots$new()}}
\item \href{#method-clone}{\code{Carrots$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Carrots$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Carrots$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
